added Middleware validator - UserController.store
