+++ src/app/controllers/UserController.js
@@ -4,20 +4,6 @@ import File from '../models/File';
 
 class UserController {
   async store(req, resp) {
-    const schema = Yup.object().shape({
-      name: Yup.string().required(),
-      email: Yup.string()
-        .email()
-        .required(),
-      password: Yup.string()
-        .required()
-        .min(6),
-    });
-
-    if (!(await schema.isValid(req.body))) {
-      return resp.status(400).json({ error: 'Validation fails' });
-    }
-
     const userExists = await User.findOne({ where: { email: req.body.email } });
 
     if (userExists) {
