+++ src/routes.js
@@ -11,12 +11,14 @@ import ScheduleController from './app/controllers/ScheduleController';
 import NotificationController from './app/controllers/NotificationController';
 import AvailabeController from './app/controllers/AvailabeController';
 
+import validateUserStore from './app/validators/UserStore';
+
 import authMiddleware from './app/middlewares/auth';
 
 const routes = new Router();
 const upload = multer(multerConfig);
 
-routes.post('/users', UserController.store);
+routes.post('/users', validateUserStore, UserController.store);
 routes.post('/sessions', SessionController.store);
 
 routes.use(authMiddleware);
